Genre Number Changed genre check to while loop to not allow user to input anything but a valid genre number. Also took out of insert song function and moved to main loop to allow repeated prompts for an appropriate genre number. Changed input method from cin to getline and atoi to fix infinite loop when user inputs a letter instead of a number. Added string temp to store int values before conversion.Putting a number in genre section now puts the song in the Alternative genre. User not prompted for another more valid genre number in this case.
Song length
Changed song length check to reduce types of inputs that cause crashes.
Function asciicheck added to Playlist. If any character in a string has a ascii value greater than 57 (that of number 9) or less than 46 (the character .) then the function returns false. Input is disregarded in main and replaced with 0 if bool add is false. This avoids the previous infinite loops.